chore: split up hir crate
