+++ Cargo.toml
@@ -2,6 +2,9 @@
 
 members = [
   "hir",
+  "hir-analysis",
+  "hir-pass",
+  "hir-type",
   "tools/*",
 ]
 
@@ -24,6 +27,8 @@ publish = false
 anyhow = "1.0"
 bitflags = "1.3"
 clap = { version = "4.1", features = ["derive"] }
+cranelift-entity = "0.95"
+cranelift-bforest = "0.95"
 env_logger = "0.9"
 log = "0.4"
 parking_lot = "0.12"
