+++ hir-analysis/src/lib.rs
@@ -8,15 +8,13 @@ pub use self::dominance::{DominanceFrontier, DominatorTree, DominatorTreePreorde
 pub use self::liveness::LivenessAnalysis;
 pub use self::loops::{Loop, LoopAnalysis, LoopLevel};
 
-use crate::hir;
-
 pub struct ControlFlowAnalysis {
     pub cfg: ControlFlowGraph,
     pub domtree: DominatorTree,
     pub loops: LoopAnalysis,
 }
 impl ControlFlowAnalysis {
-    pub fn compute(function: &hir::Function) -> Self {
+    pub fn compute(function: &miden_hir::Function) -> Self {
         let cfg = ControlFlowGraph::with_function(function);
         let domtree = DominatorTree::with_function(function, &cfg);
         let loops = LoopAnalysis::with_function(function, &cfg, &domtree);
@@ -28,11 +26,11 @@ impl ControlFlowAnalysis {
         }
     }
 
-    pub fn compute_liveness(&self, function: &hir::Function) -> LivenessAnalysis {
+    pub fn compute_liveness(&self, function: &miden_hir::Function) -> LivenessAnalysis {
         LivenessAnalysis::compute(function, self)
     }
 
-    pub fn recompute(&mut self, function: &hir::Function) {
+    pub fn recompute(&mut self, function: &miden_hir::Function) {
         self.cfg.compute(&function.dfg);
         self.domtree.compute(function, &self.cfg);
         self.loops.compute(function, &self.cfg, &self.domtree);
