+++ hir-type/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "miden-hir-type"
+version.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+repository.workspace = true
+categories.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+edition.workspace = true
+
+[dependencies]
