+++ hir/src/display.rs
@@ -1,4 +1,4 @@
-use core::{cell::Cell, fmt};
+use std::{cell::Cell, fmt};
 
 use super::{Block, Inst};
