+++ hir/src/write.rs
@@ -1,4 +1,4 @@
-use core::fmt::{self, Write};
+use std::fmt::{self, Write};
 
 use super::*;
